% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalSegments.R
\name{evalSegments}
\alias{evalSegments}
\title{Evaluate Segments}
\usage{
evalSegments(object, marker = NULL, data = c("haplotype", "dosage"))
}
\arguments{
\item{object}{A \code{genovis} class object.}

\item{marker}{A character or numeric vector specifying markers to include in the evaluation. Default is \code{NULL}.}

\item{data}{A character vector specifying which data to use, either "haplotype" or "dosage". Default is c("haplotype", "dosage").}
}
\value{
The input \code{genovis} object with evaluated segments.
}
\description{
This function evaluates the segments for haplotype and dosage data in a given \code{genovis} object.
}
