% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbscleanr2genovis.R
\name{gbscleanr2genovis}
\alias{gbscleanr2genovis}
\title{Import markers, genotypes, and haplotypes information from a GDS file}
\usage{
gbscleanr2genovis(gbsr, verbose = TRUE)
}
\arguments{
\item{gbsr}{A GbsrGenotypeData object. See \code{\link[GBScleanR:gbsrVCF2GDS]{GBScleanR::gbsrVCF2GDS()}} and
\code{\link[GBScleanR:loadGDS]{GBScleanR::loadGDS()}}.}

\item{verbose}{A logical value to indicate whether to show messages from
the function.}
}
\description{
Chromosome IDs, physical positions, and alleles of markers stored in an input
GDS file are retrieved as well as genotypes and haplotypes information at the
markers. The obtained information is then organized as a genovis class
object that can be handled by the functions in the genovisr package.
}
\examples{
# Not run
# genovis <- gdscleanr2genovis(gds_fn = "path/to/gds/file.gds",
#                              gbscleanr_filter = TRUE)

}
