% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Get Statistics Data}
\usage{
getStats(
  object,
  data = c("haplotype", "dosage", "recomb"),
  group = c("hap", "sample", "chr", "genome"),
  value = c("class", "segment")
)
}
\arguments{
\item{object}{A \code{genovis} class object.}

\item{data}{A character string specifying which data to use, either "haplotype", "dosage", or "recomb". Default is "haplotype".}

\item{group}{A character string specifying the grouping for the statistics, either "hap", "sample", "chr", or "genome". Default is "hap".}

\item{value}{A character string specifying the value to be retrieved, either "class" or "segment". Default is "class".}
}
\value{
A dataframe containing the statistics data.
}
\description{
This function retrieves statistics data for haplotype, dosage, or recombination
from a \code{genovis} object. The data can then be used for plotting graphical genotypes.
}
