% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphGeno.R
\name{plotGraphGeno}
\alias{plotGraphGeno}
\title{Plot Graphical Genotypes}
\usage{
plotGraphGeno(
  object,
  data = "haplotype",
  sample = NULL,
  direction = "h",
  margin = 0.1
)
}
\arguments{
\item{object}{A \code{genovis} class object.}

\item{data}{A character string specifying which data to use, either "haplotype" or "dosage". Default is "haplotype".}

\item{sample}{A character or numeric vector specifying samples to include in the plot. Default is \code{NULL}.}

\item{direction}{A character string specifying the plot direction, either "h" (horizontal) or "v" (vertical). Default is "h".}

\item{width}{A numeric value specifying the width of the plot elements. Default is 0.1.}
}
\value{
A ggplot object representing the graphical genotypes.
}
\description{
This function plots graphical genotypes for either haplotype or dosage data from a \code{genovis} object.
}
