% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{Plot Histogram of Segment Lengths}
\usage{
plotHist(
  object,
  data = "haplotype",
  sample = NULL,
  binwidth = 1e+06,
  fill = "darkgreen",
  chrwise = FALSE,
  samplewise = FALSE,
  ncol = NULL
)
}
\arguments{
\item{object}{A \code{genovis} class object.}

\item{data}{A character string specifying which data to use, either "haplotype" or "dosage". Default is "haplotype".}

\item{sample}{A character or numeric vector specifying samples to include in the plot. Default is \code{NULL}.}

\item{binwidth}{A numeric value specifying the bin width for the histogram. Default is 1e6 (1 Mb).}

\item{fill}{A character string specifying the fill color for the histogram bars. Default is "darkgreen".}

\item{chrwise}{A logical value indicating whether to plot histograms chromosome-wise. Default is \code{FALSE}.}

\item{samplewise}{A logical value indicating whether to plot histograms sample-wise. Default is \code{FALSE}.}

\item{ncol}{A numeric value specifying the number of columns for facet wrapping. Default is \code{NULL}.}
}
\value{
A ggplot object representing the histogram of segment lengths.
}
\description{
This function plots a histogram of segment lengths for either haplotype or dosage data from a \code{genovis} object.
}
