% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Plot Statistics for Genovis Data}
\usage{
plotStats(
  object,
  data = c("haplotype", "dosage", "recomb"),
  group = c("hap", "sample", "chr", "genome"),
  value = c("class", "segment")
)
}
\arguments{
\item{object}{A \code{genovis} class object.}

\item{data}{A character string specifying which data to use, either "haplotype", "dosage", or "recomb". Default is "haplotype".}

\item{group}{A character string specifying the grouping for the statistics, either "hap", "sample", "chr", or "genome". Default is "hap".}

\item{value}{A character string specifying the value to be plotted, either "class" or "segment". Default is "class".}
}
\value{
A ggplot object representing the statistics plot.
}
\description{
This function plots statistics for haplotype, dosage, or recombination data from a \code{genovis} object.
}
