% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print.genovis}
\alias{print.genovis}
\title{Print a summary of a genovis object}
\usage{
\method{print}{genovis}(object)
}
\arguments{
\item{object}{A \code{genovis} class object.}
}
\description{
This function print a summary of a \code{genovis} object.
}
