% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGenovir.R
\name{shinyGenovir}
\alias{shinyGenovir}
\title{Launch Shiny Application for GenovisR}
\usage{
shinyGenovir()
}
\description{
This function launches a Shiny application for visualizing and analyzing genotypic data using the GenovisR package.
}
