% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenovis.R
\name{simGenovis}
\alias{simGenovis}
\title{Create a genovis class object storing dummy data}
\usage{
simGenovis(
  n_sample = 10,
  n_marker = 1000,
  n_chr = 12,
  chr_len = sample(20:40, n_chr, replace = TRUE),
  n_ploidy = 2,
  n_haplotype = 2
)
}
\arguments{
\item{n_sample}{An integer to specify the number of samples.}

\item{n_marker}{An integer to specify the number of markers.}

\item{n_ploidy}{An integer to specify the ploidy.}

\item{n_haplotype}{An integer to specify the number of haplotypes.}
}
\description{
Create a genovis class object storing dummy data
}
\examples{
# Create simulation data with the default settings.
genovis <- simGenovis()

# Create simulation data with tweaking settings.
genovis <- simGenovis(n_sample = 100, n_amrker = 2500, n_chr = 5)

}
