% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsGeno.R
\name{statsGeno}
\alias{statsGeno}
\title{Calculate Statistics for Genovis Data}
\usage{
statsGeno(object, chr_len = NULL)
}
\arguments{
\item{object}{A \code{genovis} class object.}

\item{chr_len}{A named numeric vector specifying the lengths of each chromosome. Default is \code{NULL}, which will calculate lengths based on marker positions.}
}
\value{
A \code{genovis} class object with calculated statistics.
}
\description{
This function calculates statistics for haplotype and dosage data from a \code{genovis} object.
}
