% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbscleanr2genovis.R
\name{gbscleanr2genovis}
\alias{gbscleanr2genovis}
\title{Import markers, genotypes, and haplotypes information from a GDS file}
\usage{
gbscleanr2genovis(gds_fn, gbscleanr_filter = TRUE, ploidy = 2, verbose = TRUE)
}
\arguments{
\item{gds_fn}{A string to specify a path to the input GDS file.}

\item{gbscleanr_filter}{A logical value to indicate whether the filtering
information stored in the input GDS file should be loaded to apply the
filtering on markers and samples. See \code{\link[GBScleanR:loadGDS]{GBScleanR::loadGDS()}} and
\code{\link[GBScleanR:closeGDS]{GBScleanR::closeGDS()}}.}

\item{ploidy}{Not implemented yet.}

\item{verbose}{A logical value to indicate whether to show messages from
the function.}
}
\description{
Chromosome IDs, physical positions, and alleles of markers stored in an input
GDS file are retrieved as well as genotypes and haplotypes information at the
markers. The obtained information is then organized as a genovis class
object that can be handled by the functions in the genovisr package.
}
\examples{
# Not run
# genovis <- gdscleanr2genovis(gds_fn = "path/to/gds/file.gds",
#                              gbscleanr_filter = TRUE)

}
