% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenovis.R
\name{simGenovis}
\alias{simGenovis}
\title{Create a genovis class object storing dummy data}
\usage{
simGenovis(
  n_sample = 10,
  n_marker = 1000,
  n_chr = 12,
  chr_len = sample(20:40, n_chr, replace = TRUE),
  n_ploidy = 2,
  n_haplotype = 2
)
}
\arguments{
\item{n_sample}{An integer to specify the number of samples. Default is 10.}

\item{n_marker}{An integer to specify the number of markers. Default is 1000.}

\item{n_chr}{An integer to specify the number of chromosomes. Default is 12.}

\item{chr_len}{An integer vector to specify the length of each chromosome in megabases. Default is a random sample between 20 and 40 Mb.}

\item{n_ploidy}{An integer to specify the ploidy. Default is 2.}

\item{n_haplotype}{An integer to specify the number of haplotypes. Default is 2.}
}
\description{
Create a genovis class object storing dummy data
}
\examples{
# Create simulation data with the default settings.
genovis <- simGenovis()

# Create simulation data with tweaking settings.
genovis <- simGenovis(n_sample = 100, n_marker = 2500, n_chr = 5)

}
